+++ vite/codeSplit/src/App.jsx
@@ -5,9 +5,9 @@ import stateImg from './assets/state-mgmt.png';*/
 import { Fragment, useState } from 'react'; // Importing the useState hook from the react package to manage the state of the app component
 import Header from './components/Header'; // Importing the Header component from the Header.jsx file
 import CoreConcept from './components/CoreConcept'; // Importing the CoreConcept component from the CoreConcept.jsx file
-import TabButton from './components/TabButton'; // Importing the TabButton component from the TabButton.jsx file
+import Examples from './components/Examples';
 
-import { EXAMPLES } from './data'; // Importing the CORE_CONCEPTS array and Examples from the data.js file
+
 
 // Function to render the main content of the app with the Header and CoreConcepts components defined inline with the properties
 /*
@@ -49,43 +49,12 @@ function App() {
 // Function to render the main content of the app with the Header and CoreConcepts components defined using the CORE_CONCEPTS array
 // Using the spread operator to pass the properties of the second object in the CORE_CONCEPTS array
 function App() {
-  // selectedTopic : state variable to store the selected Topic value
-  // setSelectedTopic : function to update the selected Topic value
-  // useState('Please click a button') : initial value of the selectedTopic state variable
-  const [selectedTopic, setSelectedTopic] = useState('');
-
-  function handleSelect(selectedButton) {
-    setSelectedTopic(selectedButton);
-    console.log(selectedButton);
-  }
-  let tabContent = <p>Please select a topic</p>;
-  if (selectedTopic) {
-    tabContent = <div id="tab-content">
-      <h3>{EXAMPLES[selectedTopic].title}</h3>
-      <p>{EXAMPLES[selectedTopic].description}</p>
-      <pre>
-        <code>
-          {EXAMPLES[selectedTopic].code}
-        </code>
-      </pre>
-    </div>
-  }
-
   return (
     <Fragment>
       <Header />
       <main>
-        <CoreConcept/>
-        <section id="examples">
-          <h2>Examples</h2>
-          <menu>
-            <TabButton isSelected={selectedTopic === 'components'} onSelect={() => handleSelect('components')} >Component</TabButton>
-            <TabButton isSelected={selectedTopic === 'jsx'} onSelect={() => handleSelect('jsx')}>JSX</TabButton>
-            <TabButton isSelected={selectedTopic === 'props'} onSelect={() => handleSelect('props')}>Props</TabButton>
-            <TabButton isSelected={selectedTopic === 'state'} onSelect={() => handleSelect('state')}>State</TabButton>
-          </menu>
-          {tabContent}
-        </section>
+        <CoreConcept />
+        <Examples />
       </main>
     </Fragment>
   );
